reduce public footprint in demand.decoratedLSP.demandAgent. Reduce number of packages
