+++ src/main/java/demand/UtilityFunction.java
@@ -1,0 +1,11 @@
+package demand;
+
+import demand.demandObject.ShipperShipment;
+import demand.offer.Offer;
+
+public interface UtilityFunction {
+
+	public String getName();
+	public double getUtilityValue(Offer offer, ShipperShipment shipment);
+	
+}
