--- src/main/java/demand/utilityFunctions/UtilityFunction.java
@@ -1,11 +1,0 @@
-package demand.utilityFunctions;
-
-import demand.demandObject.ShipperShipment;
-import demand.offer.Offer;
-
-public interface UtilityFunction {
-
-	public String getName();
-	public double getUtilityValue(Offer offer, ShipperShipment shipment);
-	
-}
